requirements.txt
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
h5py>=3.8.0
scikit-learn>=1.3.0
wandb>=0.15.0
hydra-core>=1.3.0
torchmetrics>=1.0.0
neuraloperator>=1.0.0
